requirements.txt
Django>=5.2.6
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
requests>=2.31.0
django-cors-headers>=4.0.0
gunicorn>=21.2.0
whitenoise>=6.5.0
python-decouple>=3.8
dj-database-url>=2.1.0
# PostgreSQL support (optional - uncomment when ready for production)
psycopg2-binary>=2.9.7

